refactor(#1228): split workflow_check into mod + command submodule
The Finding A/B fix grew workflow_check.rs to 868 lines (>800 src threshold,
file-size ratchet FAIL). Split per the campsite rule (epic #1116): move the
pure text/shell-command parsing helpers into workflow_check/command.rs and keep
the structural jobs->steps YAML model + check_scenario orchestrator in
workflow_check/mod.rs, which re-exports the public command helpers so the
module's public API (and integration tests) are unchanged. No behavior change;
all 42 workflow tests + 100 parity tests pass, lint stays 254/0.

mod.rs 469 lines, command.rs 417 lines (both under 800).
